リファクタリング、ルート変更、albumページ追加

+++ app/Http/Controllers/NiceController.php
@@ -9,19 +9,29 @@ use Illuminate\Support\Facades\Auth;
 
 class NiceController extends Controller
 {
-    //
-    public function nice(Post $post, Request $request){
-        $nice=New Nice();
-        $nice->catimage_id=$post->id;
-        $nice->user_id=Auth::user()->id;
-        $nice->save();
-        return back();
-    }
-
-    public function unnice(Post $post, Request $request){
+    public function nice(Request $request){
         $user=Auth::user()->id;
-        $nice=Nice::where('catimage_id', $post->id)->where('user_id', $user)->first();
-        $nice->delete();
-        return back();
+        $nice=Nice::where('catimage_id', $request->id)->where('user_id', $user)->first();
+
+        if (!$nice) {
+            $nice=New Nice();
+            $nice->catimage_id=$request->id;
+            $nice->user_id=Auth::user()->id;
+            $nice->save();
+            $ohenji = "いいねしました";
+            $niced = 1;
+        } else {
+            $nice->delete();
+            $ohenji = "いいね削除しました";
+            $niced = 0;
+        }
+
+        $countNices = Nice::where('catimage_id', $request->id)->count();
+
+        return response()->json([
+            'message' => $ohenji,
+            'niced' => $niced,
+            'countNices' => $countNices,
+        ]);
     }
 }
